backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
python-multipart

